maj edit sur ensemble onglet film, série : ok niveau calcul majoritaire

+++ src/models/manga.py
@@ -4,4 +4,4 @@ from dataclasses import dataclass
 class Manga:
     id    : int = 0
     titre : str = ""
-    note  : int = 0
+    note  : float = 0
